+++ src/client/app/components/CompareBarComponent.tsx
